+++ CHANGELOG.md
@@ -16,6 +16,22 @@ Pearl changelog
 
 
 
+## [0.16.0] - 2022-12-30
+#### Added
+- Corrupted blobs count calculation (#224)
+
+#### Changed
+- Return delete or not found as distinct results (#208)
+- Save delete records in index (#207)
+- Use bytes for arguments (#191)
+
+#### Fixed
+
+
+#### Updated
+- Crate `crc` (#225)
+
+
 ## [0.15.0] - 2022-11-22
 #### Added
 - Add functions to get info about blobs and indexes (#151)